🐛 控制台最后一行输出错误

+++ src/kernel/console.c
@@ -112,7 +112,7 @@ static void scroll_up()
         screen = MEM_BASE;
     }
 
-    u32 *ptr = (u32 *)(screen + SCR_SIZE);
+    u16 *ptr = (u16 *)(screen + SCR_SIZE);
     for (size_t i = 0; i < WIDTH; i++)
     {
         *ptr++ = erase;
